Disable flake8 checks for now
FIM, which is a secondary or tertiary dependency of Fablib, pins
pycodestyle==2.6.0.

Now, pycodestyle is also a dependency of flake8, but seems that
there's no release of flake8 that can use pycodestyle==2.6.0.  All
flake8 releases either want pycodestyle>2.6.0 or pycodestyle<2.6.0.

So we'll avoid using flake8 for now, and re-enable it on CI once FIM
issue 140 is resolved.

See https://github.com/fabric-testbed/InformationModel/issues/140