Set up and Python and install black
Intention is to eventually run a series of checks on this
workflow. That may include a subset of the below, or all of them,
depending how useful they are:

 - black (https://pypi.org/project/black/)
 - isort (https://pypi.org/project/isort/)
 - flake8 (https://pypi.org/project/flake8/)
 - safety (https://pypi.org/project/safety/)
 - bandit (https://pypi.org/project/bandit/)
 - prospector (https://pypi.org/project/prospector/)

It is probably easier to install them in a Python environment, rather
than depending on their availability as a GitHub Action.