Use setuptools as the build backend
Flit won't add both fabrictestbed_fablib and fabrictestbed_extensions
into the wheel package because it is stricter about packaging.  This
is normally a good thing.  However, we want to be able to have some
backward-compatible imports at least for a while, so we will switch to
setuptools as the build backend.