Do not use an ssh agent when trying bastion connection
This is to get the honest idea on whether our bastion key setup will
work as expected, because paramiko.Client.connect() has the following
behavior:

1. first attempt to auth using pkey or key_filename argument
2. if that fails, use any keys that can be found through ssh agent
3. if that also fails, use keys from ~/.ssh

https://docs.paramiko.org/en/latest/api/client.html#paramiko.client.SSHClient.connect

We disable (3) with look_for_keys=False.  We should also disable (2)
with allow_agent=False.

This also explains why I'm getting a paramiko.SSHException (and not an
paramiko.AuthenticationException) when running integration tests on my
Debian desktop.